Add extracted CheckBox component and begin Card component and Match page refactoring
