+++ client/src/components/Card/Card.jsx
@@ -1,42 +1,31 @@
 import React from 'react'
-import { ImCheckboxUnchecked, ImCheckboxChecked } from 'react-icons/im'
 import { cardback } from '../../assets/card'
 import './Card.scss'
 
-const Card = ({ card, player, turn, handleClick, visibility, selector }) => {
+const Card = ({ card, owner, faith, handleClick, isShowing }) => {
     const { _id, power, image, values, selected } = card
 
     return (
         <div
-            className={`card ${player === 'p1' ? 'blue' : 'red'} ${
-                !turn ? 'not__user' : ''
-            } ${power} ${!visibility && 'transparent'}`}
+            className={`card ${faith === 'p1' ? 'blue' : 'red'} ${
+                !isShowing && 'transparent'
+            }`}
             id={_id}
-            owner={player}
+            owner={owner}
             onClick={(e) => handleClick(e)}
         >
             <img
                 className='card__image'
-                src={visibility ? image : cardback}
+                src={isShowing ? image : cardback}
                 alt={_id}
             />
 
-            {visibility && (
+            {isShowing && (
                 <div className='card__values'>
                     <span className='up'>{values[0]}</span>
                     <span className='right'>{values[1]}</span>
                     <span className='down'>{values[2]}</span>
                     <span className='left'>{values[3]}</span>
-                </div>
-            )}
-
-            {selector && (
-                <div className='checkbox' onClick={(e) => handleClick(e)}>
-                    {selected ? (
-                        <ImCheckboxChecked className='check' />
-                    ) : (
-                        <ImCheckboxUnchecked className='uncheck' />
-                    )}
                 </div>
             )}
         </div>
