+++ client/src/components/Card/Card.scss
@@ -130,31 +130,4 @@
             transform: translateX(-50%);
         }
     }
-
-    .checkbox {
-        position: absolute;
-        top: 5%;
-        right: 8%;
-        z-index: 2;
-
-        font-size: 32px;
-
-        .check {
-            color: var(--primary-color);
-        }
-
-        .uncheck {
-            color: var(--background-color);
-        }
-
-        svg {
-            border-radius: 7px;
-            cursor: pointer;
-        }
-
-        svg.check {
-            border: 3px solid var(--background-color);
-            font-size: 34px;
-        }
-    }
 }
